FIX: out of range condition for file transfer
git-svn-id: svn://tron.homeunix.org/simutrans/simutrans/trunk@7597 8aca7d54-2c30-db11-9de9-000461428c89

+++ network/network_file_transfer.cc
@@ -205,7 +205,7 @@ const char *network_connect(const char *cp, karte_t *world)
 			loadingscreen_t ls(translator::translate("Server preparing game ..."),300,true,true);
 #endif
 			// wait for game command for 5 min (tolerate some wrong commands) to leave it enough time for saving
-			for(uint8 i=0; i<300; i++) {
+			for(int i=0; i<300; i++) {
 #ifndef NETTOOL // no display, no translator available
 				ls.set_progress(i);
 #endif
